Code refactor KT-4
