+++ src/Application.tsx
@@ -6,7 +6,7 @@ import 'style/index.scss';
 
 export default function Application() {
     const {renderApplication} = useApplication({
-        reducers: require('./store/reducers').default,
+        reducers: require('./store').default,
         routes: () => require('routes').default,
         layoutView: () => require('shared/Layout').default,
         screen: true,
