+++ src/routes/IndexPage/views/CurrencyConversionForm/CurrencyConversionForm.tsx
@@ -3,7 +3,7 @@ import './CurrencyConversionForm.scss';
 import {useBem, useSelector} from '@steroidsjs/core/hooks';
 
 import {selectError, selectFormOne, selectFormTwo} from 'store/reducers/converter';
-import {CONVERSION_FORM_ONE, CONVERSION_FORM_TWO} from 'core/constants/currencyList';
+import {CONVERSION_FORM_ONE, CONVERSION_FORM_TWO} from 'core/constants';
 import {selectCurrencyList, selectCurrencyListIsLoading} from 'store/reducers/currency';
 import CurrencyConversion from './views/CurrencyConversion';
 
@@ -13,6 +13,8 @@ export default function CurrencyConversionForm() {
     const formOne = useSelector(selectFormOne);
     const formTwo = useSelector(selectFormTwo);
     const error = useSelector(selectError);
+
+    /** @todo Костыль. Исправить. */
     const currencyList = useSelector(selectCurrencyList);
     const isLoading = useSelector(selectCurrencyListIsLoading);
