+++ src/routes/IndexPage/views/CurrencyTable/CurrencyTable.scss
@@ -2,7 +2,6 @@
     width: 100%;
 
     &__grid {
-        border: 1px solid var(--secondary-color);
         max-width: 90%;
         margin: 0 auto;
     }
