+++ src/routes/IndexPage/views/CurrencyTable/utils.ts
@@ -1,4 +1,5 @@
 import {IGridColumn} from '@steroidsjs/core/ui/list/Grid/Grid';
+
 import {CurrencyIso} from 'core/models';
 
 interface ITable {
